image annotation complete
