--- shesha-reactjs/src/components/formDesigner/components/imageAnnotation/settingsForm.json
@@ -1,226 +1,0 @@
-{
-    "components": [
-      {
-        "id": "b8954bf6-f76d-4139-a850-c99bf06c8b69",
-        "type": "sectionSeparator",
-        "name": "separator1",
-        "parentId": "root",
-        "label": "Display"
-      },
-      {
-        "id": "5c813b1a-04c5-4658-ac0f-cbcbae6b3bd4",
-        "type": "propertyAutocomplete",
-        "name": "name",
-        "parentId": "root",
-        "label": "Name",
-        "validate": {
-          "required": true
-        }
-      },
-      {
-        "id": "46d07439-4c18-468c-89e1-60c002ce96c5",
-        "type": "textField",
-        "name": "label",
-        "parentId": "root",
-        "label": "Label"
-      },
-      {
-        "id": "57a40a33-7e08-4ce4-9f08-a34d24a83338",
-        "type": "dropdown",
-        "name": "labelAlign",
-        "parentId": "root",
-        "label": "Label align",
-        "values": [
-          {
-            "label": "left",
-            "value": "left",
-            "id": "f01e54aa-a1a4-4bd6-ba73-c395e48af8ce"
-          },
-          {
-            "label": "right",
-            "value": "right",
-            "id": "b920ef96-ae27-4a01-bfad-b5b7d07218da"
-          }
-        ],
-        "dataSourceType": "values"
-      },
-      {
-        "id": "2d32fe70-99a0-4825-ae6c-8b933004e119",
-        "type": "textArea",
-        "name": "description",
-        "parentId": "root",
-        "label": "Description"
-      },
-      {
-        "id": "df8a8f35-a50b-42f9-9642-73d390ceddbf",
-        "type": "dropdown",
-        "name": "visibility",
-        "parentId": "root",
-        "label": "Visibility",
-        "description": "This property will eventually replace the 'hidden' property and other properties that toggle visibility on the UI and payload",
-        "allowClear": true,
-        "values": [
-          {
-            "label": "Yes (Display in UI and include in payload)",
-            "value": "Yes",
-            "id": "53cd10ce-26af-474b-af75-8e7b1f19e51d"
-          },
-          {
-            "label": "No (Only include in payload)",
-            "value": "No",
-            "id": "f07a228c-cb9c-4da7-a8bc-bc2be518a058"
-          },
-          {
-            "label": "Removed (Remove from UI and exlude from payload)",
-            "value": "Removed",
-            "id": "3b6282ee-2eee-47ec-bab9-4cba52b970a0"
-          }
-        ],
-        "dataSourceType": "values"
-      },
-      {
-        "id": "cfd7d45e-c7e3-4a27-987b-dc525c412448",
-        "type": "checkbox",
-        "name": "hidden",
-        "parentId": "root",
-        "label": "Hidden"
-      },
-      {
-        "id": "c6885251-96a6-40ce-99b2-4b5209a9e01c",
-        "type": "checkbox",
-        "name": "hideLabel",
-        "parentId": "root",
-        "label": "Hide Label"
-      },
-      {
-        "id": "24a8be15-98eb-40f7-99ea-ebb602693e9c",
-        "type": "checkbox",
-        "name": "disabled",
-        "parentId": "root",
-        "label": "Disabled"
-      },
-      {
-        "id": "4b5e5951-4998-4635-b1c8-0b6d3940c300",
-        "type": "codeEditor",
-        "name": "customEnabled",
-        "label": "Custom Enabled",
-        "labelAlign": "right",
-        "parentId": "root",
-        "hidden": false,
-        "customEnabled": null,
-        "description": "Enter custom enabled code.  You must return true to enable the component. The global variable data is provided, and allows you to access the data of any form component, by using its API key.",
-        "validate": {},
-        "settingsValidationErrors": [],
-        "exposedVariables": [
-          { "name": "value", "description": "Component current value", "type": "string | any" },
-          { "name": "data", "description": "Selected form values", "type": "object" }
-        ]
-      },
-  
-      {
-        "id": "d675bfe4-ee69-431e-931b-b0e0b9ceee6f",
-        "type": "sectionSeparator",
-        "name": "separator2",
-        "parentId": "root",
-        "label": "Validation"
-      },
-      {
-        "id": "3be9da3f-f47e-48ae-b4c3-f5cc36e534d9",
-        "type": "checkbox",
-        "name": "validate.required",
-        "parentId": "root",
-        "label": "Required"
-      },
-      {
-        "id": "6befdd49-41aa-41d6-a29e-76fa00590b75",
-        "type": "sectionSeparator",
-        "name": "sectionStyle",
-        "parentId": "root",
-        "label": "Style"
-      },
-      {
-        "id": "06ab0599-914d-4d2d-875c-765a495472f8",
-        "type": "codeEditor",
-        "name": "style",
-        "label": "Style",
-        "parentId": "root",
-        "validate": {},
-        "settingsValidationErrors": [],
-        "description": "A script that returns the style of the element as an object. This should conform to CSSProperties",
-        "exposedVariables": [{ "name": "data", "description": "Form values", "type": "object" }]
-      },
-      {
-        "id": "5478b8f9-ec00-4d0a-9d2a-44a630cb2dcb",
-        "type": "sectionSeparator",
-        "name": "filesSeparator",
-        "parentId": "root",
-        "label": "Files"
-      },
-      {
-        "id": "417ee22e-a49d-44f2-a1c7-fef42ec87503",
-        "type": "textField",
-        "name": "height",
-        "parentId": "root",
-        "label": "Height"
-      },
-      {
-        "id": "c6ecd70c-7419-4ea7-a715-d42699d26e6e",
-        "type": "textField",
-        "name": "width",
-        "parentId": "root",
-        "label": "Width"
-      },
-      {
-        "id": "06ab0599-914d-4d2d-875c-765a495482f8",
-        "type": "codeEditor",
-        "name": "url",
-        "label": "Url",
-        "parentId": "root",
-        "validate": {},
-        "settingsValidationErrors": [],
-        "description": "A script that returns the image url as a string. This should return a string",
-        "exposedVariables": [{ "name": "data", "description": "Form values", "type": "object" }]
-      },
-      {
-        "id": "a59d3daa-e34f-4410-9ae1-fc1669eab5e5",
-        "type": "sectionSeparator",
-        "name": "separatorVisibility",
-        "parentId": "root",
-        "label": "Visibility"
-      },
-      {
-        "id": "fb85d916-39f9-4f77-8d87-c1c53558b078",
-        "type": "codeEditor",
-        "name": "customVisibility",
-        "label": "Custom Visibility",
-        "labelAlign": "right",
-        "parentId": "root",
-        "hidden": false,
-        "customVisibility": null,
-        "description": "Enter custom visibility code.  You must return true to show the component. The global variable data is provided, and allows you to access the data of any form component, by using its API key.",
-        "validate": {},
-        "settingsValidationErrors": [],
-        "exposedVariables": [
-          { "name": "value", "description": "Component current value", "type": "string | any" },
-          { "name": "data", "description": "Selected form values", "type": "object" }
-        ]
-      }
-    ],
-    "formSettings": {
-      "layout": "horizontal",
-      "colon": true,
-      "labelCol": {
-        "span": 5
-      },
-      "wrapperCol": {
-        "span": 13
-      },
-      "displayName": "DEFAULT_FORM_SETTINGS",
-      "__docgenInfo": {
-        "description": "Default form settings",
-        "displayName": "DEFAULT_FORM_SETTINGS",
-        "props": {}
-      }
-    }
-  }
-
